kafka/orchestrator/requirement.txt
fastapi
uvicorn
kafka-python
pydantic
python-dotenv